Thumbnail Buttons
The Buttons that I coded for the Thumbnails. I have two different versions that both work. One just needs to be activated to use the other.

Read my READ ME file, and it goes in depth about the code and what it does, and how to use both versions.

Hopefully this works, lmk if not.

I had to take out all the Images, because it wouldn't let me upload more commits with the images.